refactor: update imports and error handling in ExportTransactionsViewModel
