+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/export/ExportTransactionsViewModel.kt
@@ -16,15 +16,15 @@ import mena.wallet_presentation.generated.resources.downloading_started
 import mena.wallet_presentation.generated.resources.error
 import mena.wallet_presentation.generated.resources.error_failed_view
 import mena.wallet_presentation.generated.resources.error_no_transactions
+import mena.wallet_presentation.generated.resources.no_internet_title
 import mena.wallet_presentation.generated.resources.something_went_wrong
-import net.thechance.mena.wallet.domain.exceptions.NoDataFoundException
-import net.thechance.mena.wallet.domain.exceptions.NoInternetException
 import net.thechance.mena.wallet.domain.model.TransactionFilterParams
 import net.thechance.mena.wallet.domain.repository.StatementRepository
 import net.thechance.mena.wallet.presentation.base.BaseViewModel
-import net.thechance.mena.wallet.presentation.base.CustomToastState
-import net.thechance.mena.wallet.presentation.base.SnackBarState
+import net.thechance.mena.wallet.presentation.base.ErrorState
+import net.thechance.mena.wallet.presentation.model.CustomToastState
 import net.thechance.mena.wallet.presentation.model.FilterType
+import net.thechance.mena.wallet.presentation.model.SnackBarState
 import net.thechance.mena.wallet.presentation.screen.export.file_saver.FileSaver
 import org.jetbrains.compose.resources.StringResource
 import org.koin.android.annotation.KoinViewModel
@@ -73,17 +73,9 @@ class ExportTransactionsViewModel(
         updateState {
             newState.copy(
                 isDownloadButtonEnabled =
-                    if (newState.isCustomFilterCardSelected) {
-                        newState.hasActiveFilters
-                    } else {
-                        true
-                    },
+                    if (newState.isCustomFilterCardSelected) newState.hasActiveFilters else true,
                 isViewAndShareButtonEnabled =
-                    if (newState.isCustomFilterCardSelected) {
-                        newState.hasActiveFilters
-                    } else {
-                        true
-                    },
+                    if (newState.isCustomFilterCardSelected) newState.hasActiveFilters else true,
                 hasNoTransactionsError = false
             )
         }
@@ -138,7 +130,7 @@ class ExportTransactionsViewModel(
         sendEffect(ExportTransactionsEffect.NavigateToViewFileScreen)
     }
 
-    private suspend fun onViewAndShareError(error: Throwable) {
+    private suspend fun onViewAndShareError(error: ErrorState) {
         resetViewAndShareState()
         handleError(
             error = error,
@@ -155,10 +147,7 @@ class ExportTransactionsViewModel(
         }
 
         updateState { oldState ->
-            oldState.copy(
-                isDownloadLoading = true,
-                isViewAndShareButtonEnabled = false
-            )
+            oldState.copy(isDownloadLoading = true, isViewAndShareButtonEnabled = false)
         }
         showToast(messageRes = Res.string.downloading_started)
     }
@@ -188,7 +177,7 @@ class ExportTransactionsViewModel(
         }
     }
 
-    private suspend fun handleDownloadError(error: Throwable) {
+    private suspend fun handleDownloadError(error: ErrorState) {
         resetDownloadState()
         handleError(
             error = error,
@@ -219,10 +208,10 @@ class ExportTransactionsViewModel(
                     isSuccess = false
                 )
             }
-        } catch (error: Exception) {
+        } catch (_: Exception) {
             resetDownloadState()
             handleError(
-                error = error,
+                error = ErrorState.Unknown,
                 titleRes = Res.string.download_failed,
                 messageRes = Res.string.something_went_wrong,
                 isSuccess = false
@@ -231,13 +220,13 @@ class ExportTransactionsViewModel(
     }
 
     private suspend fun handleError(
-        error: Throwable,
+        error: ErrorState,
         titleRes: StringResource,
         messageRes: StringResource,
         isSuccess: Boolean = false
     ) {
         when (error) {
-            is NoInternetException -> {
+            is ErrorState.NoInternet -> {
                 updateState { oldState ->
                     oldState.copy(
                         noInternetConnection = true,
@@ -247,12 +236,12 @@ class ExportTransactionsViewModel(
                 }
                 showSnackBar(
                     titleRes = Res.string.download_failed,
-                    messageRes = Res.string.something_went_wrong,
+                    messageRes = Res.string.no_internet_title,
                     isSuccess = false
                 )
             }
 
-            is NoDataFoundException -> {
+            is ErrorState.NoDataFound -> {
                 updateState { oldState ->
                     oldState.copy(
                         isDownloadLoading = false,
@@ -327,19 +316,13 @@ class ExportTransactionsViewModel(
 
     private fun resetDownloadState() {
         updateState { oldState ->
-            oldState.copy(
-                isDownloadLoading = false,
-                isViewAndShareButtonEnabled = true
-            )
+            oldState.copy(isDownloadLoading = false, isViewAndShareButtonEnabled = true)
         }
     }
 
     private fun resetViewAndShareState() {
         updateState { oldState ->
-            oldState.copy(
-                isViewAndShareLoading = false,
-                isDownloadButtonEnabled = true
-            )
+            oldState.copy(isViewAndShareLoading = false, isDownloadButtonEnabled = true)
         }
     }
